% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullModels.R
\name{correct.out.data}
\alias{correct.out.data}
\title{Correct out data}
\usage{
correct.out.data(out.data.1, n.draws)
}
\arguments{
\item{out.data.1}{A data matrix with a column for every draw, and a row for every location of interest.}

\item{n.draws}{number of draws from distribution}
}
\value{
the same data matrix, but without some weird list formatting issues R was having.
}
\description{
Correct the data format coming out of the apply.X functions to be compatible
with downstream code.
}
