% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullModels.R
\name{null.engine}
\alias{null.engine}
\title{Compute the null model results}
\usage{
null.engine(this.null, train.data, n.draws, n.records, year)
}
\arguments{
\item{this.null}{an indicator for which null model to run: AA, PV, MV, HN, PN, PI, IN, UN, NB, AR}

\item{train.data}{The training data to use for parameterizing the null model}

\item{n.draws}{the number of draws from the resulting distribution to use}

\item{n.records}{The number of locations being predicted}

\item{year}{The focal forecast year}
}
\value{
a matrix with one row for each location and one column for each draw
}
\description{
Compute the null model results
}
